new_backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
apscheduler
python-dotenv
pandas
yfinance
numpy
joblib
scikit-learn
xgboost
gunicorn